requirements.txt
aiofiles==24.1.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jwt==1.4.0
ldap3==2.9.1
Mako==1.3.10
MarkupSafe==3.0.2
paramiko==3.5.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyad==0.6.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.1.0
python-jose==3.4.0
python-ldap==3.4.3
python-multipart==0.0.20
requests==2.31.0
rsa==4.9.1
schedule==1.2.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
Werkzeug==3.1.3
